requirements.txt
certifi==2020.6.20
chardet==3.0.4
docker==4.2.2
idna==2.10
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
websocket-client==0.57.0
